[*] signal.hpp to signal_callback.hpp/cpp
