+++ CMakeLists.txt
@@ -27,7 +27,8 @@ add_library(co_lib
         src/co/net/tcp_listener.cpp
         src/co/net/impl/uv_tcp_ptr.cpp
         src/co/net/tcp_stream.cpp
-        src/co/condition_variable.cpp)
+        src/co/condition_variable.cpp
+        src/co/signal_callback.cpp)
 
 target_link_libraries(co_lib PUBLIC ${CONAN_LIBS})
 target_link_directories(co_lib PUBLIC ${CONAN_LIB_DIRS})
